backend/requirements.txt
selenium
bs4
webdriver-manager
requests
python-dotenv
subprocess.run 
